+++ backend/task-manager/src/test/java/com/example/task_manager/controller_tests/IsAssignedControllerTest.java
@@ -8,7 +8,6 @@ import com.example.task_manager.DTO.IsAssignedDTO;
 import com.example.task_manager.controller.IsAssignedController;
 import com.example.task_manager.service.IsAssignedService;
 import org.junit.jupiter.api.Test;
-import org.mockito.InjectMocks;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
@@ -24,9 +23,6 @@ public class IsAssignedControllerTest {
 
     @MockBean
     private IsAssignedService isAssignedService;
-
-    @InjectMocks
-    private IsAssignedController isAssignedController;
 
     /**
      * Assign a Team Member to a Task
